refactor(routes): extract settings routes
